Correctly place buttons added after controllers.
This feels pretty hacky -- not only because it necessitates moving
back to "all button position values are pixel values" -- but because
the UI manager now has to ask the controller manager if the button
being added is a controller button. That makes it very chatty 2-way
communication between these two and I don't like it.

The alternative is to add a piece of data to the button that says it
should ignore the canvas' top/left offsets, which .. is likely a
better solution. Still, this is what I've done for now and I'll have
to revisit it soon anyway.